+++ src/__tests__/built/pipe.test.ts
@@ -1,6 +1,6 @@
 import { THIS1 } from '@bemedev/dev-utils/build-tests';
 import { createTests } from '@bemedev/dev-utils/vitest-extended';
-import { add, division, exponent, times } from '../../extensions';
+import { addBy, divisionBy, exponentBy, timesBy } from '../../extensions';
 
 describe('All imports', () => {
   describe('#01 => pipe exists at paths', () => {
@@ -22,49 +22,54 @@ describe('All imports', () => {
       success(
         {
           invite: '(2+1)*2 = 6',
-          parameters: [add(1), times(2)] as any,
+          parameters: [addBy(1), timesBy(2)] as any,
           expected: 6,
         },
         {
           invite: '((2+1)*2)-3 = 3',
-          parameters: [add(1), times(2), add(-3)] as any,
+          parameters: [addBy(1), timesBy(2), addBy(-3)] as any,
           expected: 3,
         },
         {
           invite: '(((2+1)*2)-3)/2 = 1.5',
-          parameters: [add(1), times(2), add(-3), division(2)] as any,
+          parameters: [
+            addBy(1),
+            timesBy(2),
+            addBy(-3),
+            divisionBy(2),
+          ] as any,
           expected: 1.5,
         },
         {
           invite: '((((2+1)*2)-3)/2)^2 = 2.25',
           parameters: [
-            add(1),
-            times(2),
-            add(-3),
-            division(2),
-            exponent(2),
+            addBy(1),
+            timesBy(2),
+            addBy(-3),
+            divisionBy(2),
+            exponentBy(2),
           ] as any,
           expected: 2.25,
         },
         {
           invite: '((((((((2+1)*2)-3)/2)^2)+10)/2)-1)*3/5 = 3.075',
           parameters: [
-            add(1),
-            times(2),
-            add(-3),
-            division(2),
-            exponent(2),
-            add(10),
-            division(2),
-            add(-1),
-            times(3),
-            division(5),
+            addBy(1),
+            timesBy(2),
+            addBy(-3),
+            divisionBy(2),
+            exponentBy(2),
+            addBy(10),
+            divisionBy(2),
+            addBy(-1),
+            timesBy(3),
+            divisionBy(5),
           ] as any,
           expected: 3.075,
         },
         {
-          invite: 'Add 1 20 times',
-          parameters: Array.from({ length: 20 }, () => add(1)) as any,
+          invite: 'Add 1 20 timesBy',
+          parameters: Array.from({ length: 20 }, () => addBy(1)) as any,
           expected: 22,
         },
       ),
@@ -78,21 +83,21 @@ describe('All imports', () => {
       success(
         {
           invite: '(2+1)*2 = 6',
-          parameters: [addAsync(1), times(2)] as any,
+          parameters: [addAsync(1), timesBy(2)] as any,
           expected: 6,
         },
         {
           invite: '((2+1)*2)-3 = 3',
-          parameters: [addAsync(1), times(2), addAsync(-3)] as any,
+          parameters: [addAsync(1), timesBy(2), addAsync(-3)] as any,
           expected: 3,
         },
         {
           invite: '(((2+1)*2)-3)/2 = 1.5',
           parameters: [
             addAsync(1),
-            times(2),
+            timesBy(2),
             addAsync(-3),
-            division(2),
+            divisionBy(2),
           ] as any,
           expected: 1.5,
         },
@@ -100,10 +105,10 @@ describe('All imports', () => {
           invite: '((((2+1)*2)-3)/2)^2 = 2.25',
           parameters: [
             addAsync(1),
-            times(2),
+            timesBy(2),
             addAsync(-3),
-            division(2),
-            exponent(2),
+            divisionBy(2),
+            exponentBy(2),
           ] as any,
           expected: 2.25,
         },
@@ -111,25 +116,25 @@ describe('All imports', () => {
           invite: '((((((((2+1)*2)-3)/2)^2)+10)/2)-1)*3/5 = 3.075',
           parameters: [
             addAsync(1),
-            times(2),
+            timesBy(2),
             addAsync(-3),
-            division(2),
-            exponent(2),
+            divisionBy(2),
+            exponentBy(2),
             addAsync(10),
-            division(2),
+            divisionBy(2),
             addAsync(-1),
-            times(3),
-            division(5),
+            timesBy(3),
+            divisionBy(5),
           ] as any,
           expected: 3.075,
         },
         {
-          invite: 'AddAsync 1 (20 times)',
+          invite: 'AddAsync 1 (20 timesBy)',
           parameters: Array.from({ length: 20 }, () => addAsync(1)) as any,
           expected: 22,
         },
         {
-          invite: 'AddAsync 1 (1000 times)',
+          invite: 'AddAsync 1 (1000 timesBy)',
           parameters: Array.from({ length: 1000 }, () =>
             addAsync(1),
           ) as any,